Intégration du composant MenuScrollTo et importation dans app.js
